Extract: matchObject to own file
