+++ lib/create-matcher.js
@@ -16,10 +16,10 @@ var assertMethodExists = require("./create-matcher/assert-method-exists");
 var assertType = require("./create-matcher/assert-type");
 var isIterable = require("./create-matcher/is-iterable");
 var isMatcher = require("./create-matcher/is-matcher");
+var matchObject = require("./create-matcher/match-object");
 var matcherPrototype = require("./create-matcher/matcher-prototype");
 
 var arrayIndexOf = arrayProto.indexOf;
-var arrayEvery = arrayProto.every;
 var join = arrayProto.join;
 var map = arrayProto.map;
 var some = arrayProto.some;
@@ -28,39 +28,6 @@ var hasOwnProperty = objectProto.hasOwnProperty;
 var objectToString = objectProto.toString;
 
 var stringIndexOf = stringProto.indexOf;
-
-/**
- * Matches `actual` with `expectation`
- *
- * @private
- * @param {*} actual A value to examine
- * @param {object} expectation An object with properties to match on
- * @returns {boolean} Returns true when `actual` matches all properties in `expectation`
- */
-function matchObject(actual, expectation) {
-    if (actual === null || actual === undefined) {
-        return false;
-    }
-
-    return arrayEvery(Object.keys(expectation), function(key) {
-        var exp = expectation[key];
-        var act = actual[key];
-
-        if (isMatcher(exp)) {
-            if (!exp.test(act)) {
-                return false;
-            }
-        } else if (typeOf(exp) === "object") {
-            if (!matchObject(act, exp)) {
-                return false;
-            }
-        } else if (!deepEqual(act, exp)) {
-            return false;
-        }
-
-        return true;
-    });
-}
 
 var TYPE_MAP = {
     function: function(m, expectation, message) {
